inBounds overhaul
Changed the way that inBounds, isThrough and nodesOutOfBounds work,
before they just checked if all inside nodes were inside corridor, but
that is not correct. Only the part of the inside nodes that are actually
in the shape should be checked.
They all now rely on arc-line and line-line intersections instead